fixed user registration
